Split linear algebra tests into separate files in their own folder
